+++ worker-configuration.d.ts
@@ -1,10 +1,7 @@
 /* eslint-disable */
-// Generated by Wrangler by running `wrangler types` (hash: 9d50a76cdc759facba0bb43994cbf977)
+// Generated by Wrangler by running `wrangler types` (hash: 187132f48ddf0f604882ba8213fe386f)
 // Runtime types generated with workerd@1.20251118.0 2025-08-15 nodejs_compat
 declare namespace Cloudflare {
-	interface GlobalProps {
-		mainModule: typeof import("./dist/_worker.js/index");
-	}
 	interface Env {
 		ASSETS: Fetcher;
 	}
